requirements.txt
# CS 412 Research Project Requirements

## Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
tensorflow>=2.8.0
torch>=1.11.0
transformers>=4.20.0

## Data Processing
json5>=0.9.0
pyarrow>=7.0.0
h5py>=3.6.0

## Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0

## Text Processing
nltk>=3.7
spacy>=3.4.0
textblob>=0.17.0

## Utilities
tqdm>=4.64.0
joblib>=1.1.0
optuna>=3.0.0
imbalanced-learn>=0.9.0

## Development
jupyter>=1.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

## Configuration
pyyaml>=6.0
python-dotenv>=0.19.0
